mkLTSAsessions batch processing option
Main changes:

- Turn mkLTSAsessions into a function (OPTIONAL!) If you call it the
usual way, with no arguments, it prompts you to enter all the file and
species info as usual.
BUT, you can pass those things in as arguments instead if you want to.

- Include a new function **itr_mkLTSA.m** which can call mkLTSA sessions
for all TPWS files in a directory, for batch processing.

- Turn sTime, eTime, and rTime (LTSA header info variables) into
"Persistent" variables, so if you are using batch processing the headers
are only read once. This is for speed.

Other changes:
- Add calculation of raw file duration from LTSA Params, instead of
assuming it's always 75 sec. This is for non-200 kHz deployments.