Multiple things
findTffile - tf for DT01 was wrong, updated for the new tf numbers

SumPPICIBin it is set to go. It is the previous step for doing Density. It produces a mat file summaryData_forDensity, which has the parameters per click and bin and effort times.

The other files is for Pm, which create the effort times based on the ship times and general effort.