calling detEdit simplified
I did some changes on how detEdit is called now, just to make it easier for the user to be able to change settings and run the code.
So all the parameter defaults are organized in initDefaultParams. The species parameters are organized in initSpParams.
Then you can create your own script and put all the directories and temporal changes that you would like to run (that overwrites default and species default parameters).
Now you don't have to modify detEdit to call your settings. You can do it just by typing detEdit(@yourDataSettings) or if you just type detEdit, it will ask to select the script that you want.
I already created the setting files that we had from all of us, like dolphinSettings, spermwhaleSettings, etc. So in this case you can call detEdit like detEdit(@spermwhaleSettings).

I haven't changed it for mkLTSA and modDet, just want to hear from you if you like it, so I will go ahead and will do all the changes that need to be done.