fix small bug
as far as I can tell, pt and pwr weren't clearing in between files. So if disk01a_LTSA1 had 30 sessions, anything after that with <30 sessions would end up with 30 sessions anyway. This caused issues when opening files in detEdit, because the calculated number of LTSA sessions didn't match what was stored in the LTSA1 files.